requirements.txt
python-dotenv>=1.0.0
fastapi
uvicorn[standard]
websockets
motor
pymongo
redis[hiredis]>=5.0.0
